Break out the instruction enum into its own file
