+++ core/bytecode/src/lib.rs
@@ -5,6 +5,7 @@
 
 mod chunk;
 mod compiler;
+mod instruction;
 mod instruction_reader;
 mod loader;
 mod op;
@@ -12,7 +13,8 @@ mod op;
 pub use crate::{
     chunk::{Chunk, DebugInfo},
     compiler::{Compiler, CompilerError, CompilerSettings},
-    instruction_reader::{FunctionFlags, Instruction, InstructionReader, TypeId},
+    instruction::{FunctionFlags, Instruction, TypeId},
+    instruction_reader::InstructionReader,
     loader::{Loader, LoaderError},
     op::Op,
 };
